Fix reproducibility of non-uclibc i386 builds
Without this fix, if we try to build i386 on an amd64 machine without the linux32 "personality" applied, it uses `uname -m` and gets a slightly different result due to the lack of https://git.busybox.net/busybox/tree/arch/i386/Makefile?h=1_37_stable#n7 (adding `-march` and `-mtune` values to builds).

If we set `ARCH` appropriately during build, that file gets sourced and our builds include the appropriate values.

This does *not* affect uclibc builds because they set `CROSS_COMPILE`, which overrides the value of `(SUB)ARCH` (although it would be valid to set both in all cases, to be explicit, I just don't think that's necessary).