Drop support for mips64le
With the release of Debian Trixie, this is no longer a "release architecture" in Debian, so our ability (and frankly desire) to support it here is dry.

https://www.debian.org/releases/trixie/release-notes/issues.en.html#mips-architectures-removed

> From trixie, the architectures `mipsel` and `mips64el` are no longer supported by Debian. Users of these architectures are advised to switch to different hardware.

I intentionally left the `mips`-related bits in `Dockerfile-builder.template` as they're still technically valid and it felt bad to remove that knowledge right now (they'll be totally unused and are trivial to remove later but harder to dig back up and re-apply).

In theory, we could build on Debian Unstable instead for `mips64le`, but that's something that likely won't continue to be true for very long unless there's a larger effort to revitalize the architecture.

We could also consider cross-compiling, but at some point we're chasing a ghost and should just cut bait.